updated stuff

+++ src/web-app/server/Code.cs
@@ -7,7 +7,7 @@ namespace CodeJar.WebApp
     public class Code
     {
         public int ID { get; set; }
-        public long SeedValue { get; set; }
+        public int SeedValue { get; set; }
         public string StringValue { get; set; }
         public string State { get; set; }
         public DateTime DateActive { get; set; }
